split into datatypes
