+++ data/data_suite_test.go
@@ -1,8 +1,8 @@
 package data_test
 
 import (
-	. "github.com/tidepool-org/platform/Godeps/_workspace/src/github.com/onsi/ginkgo"
-	. "github.com/tidepool-org/platform/Godeps/_workspace/src/github.com/onsi/gomega"
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
 
 	"testing"
 )
